Implement changes requested:
If the Preferred Language is not English  and doesn't have a constant configured (so there is no default message for that language), please use the English one by default and in the pop-up update the successful message with the following: The translated Welcome SMS for <language> was unavailable, so the English version was used. The Welcome SMS was sent successfully

If English is not available either, then show the following error message:
The Welcome SMS could not be sent because there is not a valid Welcome SMS template configured.

If possible, please move 'Send SMS' option to the top of the menu, above Help.

Please, change 'Send SMS' for 'Send Welcome SMS'. Likewise, update the default OK/KO error messages (the ones you use after getting an OK/KO from the gateway) to say 'Welcome SMS' instead of just 'SMS':
There was an error sending the Welcome SMS

The Welcome SMS was sent sucessfully

for the time being, could you please disable the automatic SMS sending? So it is only sent if you use the option in the menu.

The client would like to test it with the SMS being only sent if you use the option in the menu, but since we've already prepared the automatic SMS sending after the enrollment is created, please leave it disabled just in case we need it in the future.

Thanks!